+++ world/src/all.rs
@@ -19,6 +19,9 @@ pub enum ForestKind {
     Swamp,
     Frostpine,
     Dead,
+    Mapletree,
+    Cherry,
+    AutumnTree,
 }
 
 pub struct Environment {
@@ -43,6 +46,9 @@ impl ForestKind {
             ForestKind::Mangrove => 0.5..1.3,
             ForestKind::Swamp => 0.5..1.1,
             ForestKind::Dead => 0.0..1.5,
+            ForestKind::Mapletree => 0.65..1.25,
+            ForestKind::Cherry => 0.45..0.75,
+            ForestKind::AutumnTree => 0.25..0.65,
             _ => 0.0..0.0,
         }
     }
@@ -62,6 +68,9 @@ impl ForestKind {
             ForestKind::Mangrove => 0.35..1.6,
             ForestKind::Swamp => -0.6..0.8,
             ForestKind::Dead => -1.5..1.0,
+            ForestKind::Mapletree => -0.15..0.25,
+            ForestKind::Cherry => -0.10..0.15,
+            ForestKind::AutumnTree => -0.45..0.05,
             _ => 0.0..0.0,
         }
     }
@@ -90,6 +99,9 @@ impl ForestKind {
             ForestKind::Mangrove => 2.0,
             ForestKind::Swamp => 1.0,
             ForestKind::Dead => 0.01,
+            ForestKind::Mapletree => 0.65,
+            ForestKind::Cherry => 12.0,
+            ForestKind::AutumnTree => 125.0,
             _ => 0.0,
         }
     }
@@ -107,6 +119,9 @@ impl ForestKind {
             ForestKind::Birch => 0.65,
             ForestKind::Mangrove => 1.0,
             ForestKind::Swamp => 0.4,
+            ForestKind::Mapletree => 0.4,
+            ForestKind::Cherry => 0.3,
+            ForestKind::AutumnTree => 0.4,
             _ => 1.0,
         }
     }